requirements.txt
asgiref==3.8.1
Django==4.2.20
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.0
